requirements.txt
Django==1.5.5
South==0.8.2
django-crispy-forms==1.4.0
python-dateutil==2.2